+++ app/derivative_services/jp2_derivative_service.rb
@@ -55,7 +55,7 @@ class Jp2DerivativeService
   end
 
   def valid_mime_types
-    ["image/tiff", "image/jpeg", "image/png", "image/jp2"]
+    ["image/tiff", "image/jpeg", "image/png"]
   end
 
   def create_derivatives
@@ -79,11 +79,7 @@ class Jp2DerivativeService
   end
 
   def run_tiff_derivatives
-    if mime_type.first == "image/jp2"
-      copy_file(filename)
-    else
-      create_tiff_derivative(filename)
-    end
+    create_tiff_derivative(filename)
   end
 
   def create_tiff_derivative(filename)
@@ -91,10 +87,6 @@ class Jp2DerivativeService
     _stdout, stderr, status =
       Open3.capture3("opj_compress", "-i", color_corrected_tiff.path.to_s, "-o", temporary_output.path.to_s, "-t", "1024,1024", "-p", "RPCL", "-n", "8", "-r", "10")
     raise stderr unless status.success?
-  end
-
-  def copy_file(filename)
-    FileUtils.cp(filename.to_s, temporary_output.path.to_s)
   end
 
   def correct_color(filename)