+++ spec/derivative_services/jp2_derivative_service_spec.rb
@@ -48,15 +48,6 @@ RSpec.describe Jp2DerivativeService do
       end
     end
 
-    context "when given a JPEG2000 mime_type" do
-      it "is valid" do
-        # rubocop:disable RSpec/SubjectStub
-        allow(valid_file).to receive(:mime_type).and_return(["image/jp2"])
-        # rubocop:enable RSpec/SubjectStub
-        is_expected.to be_valid
-      end
-    end
-
     context "when given an invalid mime_type" do
       it "does not validate" do
         # rubocop:disable RSpec/SubjectStub
@@ -160,20 +151,6 @@ RSpec.describe Jp2DerivativeService do
   context "png source", run_real_derivatives: true do
     let(:file) { fixture_file_upload("files/abstract.png", "image/png") }
     it "creates a JP2 and attaches it to the fileset" do
-      derivative_service.new(id: valid_change_set.id).create_derivatives
-
-      reloaded = query_service.find_by(id: valid_resource.id)
-      derivative = reloaded.derivative_file
-
-      expect(derivative).to be_present
-      derivative_file = Valkyrie::StorageAdapter.find_by(id: derivative.file_identifiers.first)
-      expect(derivative_file.read).not_to be_blank
-    end
-  end
-
-  context "JPEG2000 source", run_real_derivatives: true do
-    let(:file) { fixture_file_upload("files/example.jp2", "image/jp2") }
-    it "creates a derivative JP2 and attaches it to the fileset" do
       derivative_service.new(id: valid_change_set.id).create_derivatives
 
       reloaded = query_service.find_by(id: valid_resource.id)